+++ src/core/defaultOptions.ts
