+++ src/core/eventListenerManager.ts
