+++ src/index.ts
@@ -1,2 +1,2 @@
 export type { DarkifyPlugin, Options } from '@/types';
-export { Darkify as default } from '@/darkify';
+export { Darkify as default } from '@/core/darkify';
