+++ src/plugins/index.ts
@@ -1,5 +1,2 @@
-export { ThemeWidget } from './themeWidget';
-export type { ThemeWidgetOptions } from './themeWidget';
-
-export { KeyboardShortcut } from './keyboardShortcut';
-export type { KeyboardShortcutOptions } from './keyboardShortcut';
+export { ThemeWidget } from './widget';
+export { KeyboardShortcut } from './keyboard';
