+++ src/types/index.d.ts
@@ -1,13 +1,15 @@
-export interface DarkifyPlugin {
-  el?: HTMLElement | ShadowRoot;
-  render(): void | HTMLElement | ShadowRoot;
+import type { LitElement } from 'lit';
+
+export interface DarkifyPlugin<T extends LitElement = LitElement> {
+  el?: T;
+  render(): void | T;
   onThemeChange?: (theme: string) => void;
   onDestroy?: () => void;
 }
 
-export interface DarkifyPluginElement {
-  new (host: any, options?: object): DarkifyPlugin;
+export interface DarkifyPluginConstructor {
   pluginId: string;
+  new (host: any, options?: any): DarkifyPlugin;
 }
 
 export type StorageType = 'local' | 'session' | 'none';
@@ -16,5 +18,5 @@ export interface Options {
   autoMatchTheme: boolean;
   useColorScheme: [string, string?];
   useStorage: StorageType;
-  usePlugins?: (DarkifyPluginElement | [DarkifyPluginElement, any])[];
+  usePlugins?: (DarkifyPluginConstructor | [DarkifyPluginConstructor, any])[];
 }
